+++ tests/config.rs
@@ -1,4 +1,4 @@
-use dook::config::Config;
+use dook::config::{Config, QueryCompiler};
 use dook::language_name::LanguageName;
 use dook::loader;
 
@@ -10,8 +10,9 @@ fn default_patterns_are_loadable() {
     let mut language_loader =
         loader::Loader::new(target_dir.clone(), Some(target_dir.clone()), false);
     let default_config = Config::load_default();
+    let mut query_compiler = QueryCompiler::new(&default_config);
     for language_name in LanguageName::iter() {
-        default_config
+        query_compiler
             .get_language_info(language_name, &mut language_loader)
             .unwrap()
             .unwrap();